refactor all snake yaml usages
created defaults for new signatures
