+++ server/src/main/java/org/cloudfoundry/identity/uaa/impl/config/CustomPropertyConstructor.java
@@ -12,9 +12,8 @@
  *******************************************************************************/
 package org.cloudfoundry.identity.uaa.impl.config;
 
+import org.cloudfoundry.identity.uaa.util.UaaYamlUtils;
 import org.springframework.util.Assert;
-import org.yaml.snakeyaml.DumperOptions;
-import org.yaml.snakeyaml.LoaderOptions;
 import org.yaml.snakeyaml.TypeDescription;
 import org.yaml.snakeyaml.constructor.Constructor;
 import org.yaml.snakeyaml.introspector.Property;
@@ -37,7 +36,7 @@ public class CustomPropertyConstructor extends Constructor {
     private final Map<Class<?>, AliasSupportingTypeDescription> typeDescriptions = new HashMap<>();
 
     public CustomPropertyConstructor(Class<?> theRoot) {
-        super(theRoot, getDefaultLoaderOptions());
+        super(theRoot, UaaYamlUtils.getDefaultLoaderOptions());
         TypeDescription typeDescription = createTypeDescription(theRoot);
         addTypeDescription(typeDescription);
         yamlClassConstructors.put(NodeId.mapping, new CustomPropertyConstructMapping());
@@ -48,22 +47,6 @@ public class CustomPropertyConstructor extends Constructor {
         AliasSupportingTypeDescription typeDescription = new AliasSupportingTypeDescription(clazz);
         typeDescriptions.put(clazz, typeDescription);
         return typeDescription;
-    }
-
-    public static LoaderOptions getDefaultLoaderOptions() {
-        LoaderOptions loaderOptions = new LoaderOptions();
-        loaderOptions.setAllowDuplicateKeys(false);
-        return loaderOptions;
-    }
-
-    public static DumperOptions getDefaultDumperOptions() {
-        DumperOptions dump = new DumperOptions();
-        dump.setDefaultFlowStyle(DumperOptions.FlowStyle.BLOCK);
-        dump.setPrettyFlow(true);
-        dump.setIndent(2);
-        dump.setCanonical(false);
-        dump.setExplicitStart(true);
-        return dump;
     }
 
     /**
